backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
requests==2.31.0
python-multipart==0.0.6
redis==5.0.1
aiohttp==3.9.1
loguru==0.7.2
alembic==1.13.1
